Abstract the selectProofsToSend algo, adds tests for edge cases, removes redundant coinselection tests from wallet (as most branches covered by wallet send tests)
