+++ src/wallet/index.ts
@@ -1,6 +1,7 @@
 export * from './counters';
 export * from './KeyChain';
 export * from './Keyset';
+export * from './selectProofsRGLI';
 export * from './types';
 export * from './Wallet';
 export * from './WalletEvents';
